fix: installed the correct version of victory-native
