+++ screens/DataViz.js
@@ -11,35 +11,34 @@ import { db } from '../config/firebase';
 // import { dogsData } from '../src/data/data';
 
 export default function App(props) {
-//    const [allDogs, setAllDogs] = React.useState([]);
-//    React.useEffect(() => {
-//       // dogs database (allDogs)
-//       const unsubscribe = db.collectionGroup('dogs').onSnapshot(
-//          snapshot => {
-//             const allDogsArr = [];
-//             snapshot.forEach(doc => {
-//                const { breed, imageUrl, location, points } = doc.data();
-//                allDogsArr.push({
-//                   key: doc.id,
-//                   breed,
-//                   imageUrl,
-//                   location,
-//                   points
-//                });
-//                // console.log('hello', doc.id, ' => ', doc.data());
-//             });
-//             // setLoading(false);
-//             setAllDogs(allDogsArr);
-//          }
-//          //  err => {
-//          //     setError(err);
-//          //  }
-//       );
-//       return () => unsubscribe();
-//    }, []);
-//    console.log('hilary', allDogs);
+   const [allDogs, setAllDogs] = React.useState([]);
+   React.useEffect(() => {
+      // dogs database (allDogs)
+      const unsubscribe = db.collectionGroup('dogs').onSnapshot(
+         snapshot => {
+            const allDogsArr = [];
+            snapshot.forEach(doc => {
+               const { breed, imageUrl, location, points } = doc.data();
+               allDogsArr.push({
+                  key: doc.id,
+                  breed,
+                  imageUrl,
+                  location,
+                  points
+               });
+               // console.log('hello', doc.id, ' => ', doc.data());
+            });
+            // setLoading(false);
+            setAllDogs(allDogsArr);
+         }
+         //  err => {
+         //     setError(err);
+         //  }
+      );
+      return () => unsubscribe();
+   }, []);
+   console.log('allDogs', allDogs);
 
-   //    render() {
    return (
       <ScrollView>
          <VictoryChart>
@@ -57,7 +56,6 @@ export default function App(props) {
             </VictoryChart> */}
       </ScrollView>
    );
-   //    }
 }
 
 const styles = StyleSheet.create({
