+++ src/data/data.js
@@ -1,29 +1,4 @@
-import { db } from '../../config/firebase';
-
-
-export const dogsData = 
-db.collectionGroup('dogs').onSnapshot(
-   snapshot => {
-      const allDogsArr = [];
-      snapshot.forEach(doc => {
-         const { breed, imageUrl, location, points } = doc.data();
-         allDogsArr.push({
-            key: doc.id,
-            breed,
-            imageUrl,
-            location,
-            points
-         });
-         // console.log('hello', doc.id, ' => ', doc.data());
-      });
-      // setLoading(false);
-      return allDogsArr
-    //   setAllDogs(allDogsArr);
-   }
-   //  err => {
-   //     setError(err);
-   //  }
-);
+// import { db } from '../../config/firebase'
 
 export const data = [
    { quarter: 1, earnings: 13000 },
